requirements.txt
torch==2.4.0
dgl==2.1.0
emoji
numpy
pandas
munkres
sentence_transformers
openpyxl
matplotlib
